requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-groq
cryptography
sqlalchemy
python-dotenv
requests
bcrypt
plotly
pandas
slowapi
psycopg2-binary
openpyxl
PyPDF2
python-multipart
python-docx
httpx
beautifulsoup4
apscheduler